+++ CHANGELOG.md
@@ -21,6 +21,7 @@ v2.0.0-beta.10
  * ons-page: Update API (`disabled`, `backButtonHandler` properties).
  * ons-button, ons-ripple: Update API (`disabled` property).
  * ons-speeddial: Updated API (`disabled`, `visible`, `inline` properties).
+ * ons-switch: Updated API (`disabled`, `checked`, `checkbox` properties).
 
 v2.0.0-beta.9
 ----